% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterDataParser.R
\name{clusterDataParser}
\alias{clusterDataParser}
\title{clusterDataParser}
\description{
\code{clusterDataParser} reads in UC format clustering infor and makes convenient
data strucrures for cluster merging and representative - cluster members
mapping.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{df}}{(\code{data.table})\cr
Stores clustering data - two columns - representative and members}

\item{\code{seq_rep}}{(\code{list})\cr
names list maping a seqence ID to the reresentative  ID}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clustersclusterDataParser-new}{\code{clusterDataParser$new()}}
\item \href{#method-clustersclusterDataParser-get_members}{\code{clusterDataParser$get_members()}}
\item \href{#method-clustersclusterDataParser-clone}{\code{clusterDataParser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clustersclusterDataParser-new"></a>}}
\if{latex}{\out{\hypertarget{method-clustersclusterDataParser-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clusterDataParser$new(fileName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileName}}{(\code{character(1)})\cr
An UC file (like those produced by Vsearch tools).
Read in the UC file and parse the produced df}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clustersclusterDataParser-get_members"></a>}}
\if{latex}{\out{\hypertarget{method-clustersclusterDataParser-get_members}{}}}
\subsection{Method \code{get_members()}}{
Returns a list of cluster menbers IDs given the  representative ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clusterDataParser$get_members(repr_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repr_id}}{(\code{character(1)})\cr
A ccluster representative sequence id}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clustersclusterDataParser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clustersclusterDataParser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clusterDataParser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
